Remove product form actions factory unit tests when Products M3 is disabled.
