+++ WooCommerce/WooCommerceTests/ViewRelated/Products/BottomSheetListSelector/ProductFormActionsFactory+NonEmptyBottomSheetActionsTests.swift
@@ -3,65 +3,6 @@ import XCTest
 @testable import Yosemite
 
 final class ProductFormActionsFactory_NonEmptyBottomSheetActionsTests: XCTestCase {
-
-    // M3 feature flag off & M2 feature flag on
-
-    func testDataHasNoEditProductsRelease3ActionsForAPhysicalProductWhenM3FeatureFlagIsOff() {
-        // Arrange
-        let product = Fixtures.physicalProduct
-        let model = EditableProductModel(product: product)
-
-        // Action
-        let factory = ProductFormActionsFactory(product: model,
-                                                formType: .edit,
-                                                isEditProductsRelease3Enabled: false)
-
-        // Assert
-        let expectedSettingsSectionActions: [ProductFormEditAction] = [.priceSettings]
-        XCTAssertEqual(factory.settingsSectionActions(), expectedSettingsSectionActions)
-
-        let expectedBottomSheetActions: [ProductFormBottomSheetAction] = [.editShippingSettings, .editInventorySettings, .editBriefDescription]
-        XCTAssertEqual(factory.bottomSheetActions(), expectedBottomSheetActions)
-    }
-
-    func testDataHasNoEditProductsRelease3AndShippingActionsForAVirtualProductWhenM3FeatureFlagIsOff() {
-        // Arrange
-        let product = Fixtures.virtualProduct
-        let model = EditableProductModel(product: product)
-
-        // Action
-        let factory = ProductFormActionsFactory(product: model,
-                                                formType: .edit,
-                                                isEditProductsRelease3Enabled: false)
-
-        // Assert
-        let expectedSettingsSectionActions: [ProductFormEditAction] = [.priceSettings]
-        XCTAssertEqual(factory.settingsSectionActions(), expectedSettingsSectionActions)
-
-        let expectedBottomSheetActions: [ProductFormBottomSheetAction] = [.editInventorySettings, .editBriefDescription]
-        XCTAssertEqual(factory.bottomSheetActions(), expectedBottomSheetActions)
-    }
-
-    func testDataHasNoEditProductsRelease3AndShippingActionsForADownloadableProductWhenM3FeatureFlagIsOff() {
-        // Arrange
-        let product = Fixtures.downloadableProduct
-        let model = EditableProductModel(product: product)
-
-        // Action
-        let factory = ProductFormActionsFactory(product: model,
-                                                formType: .edit,
-                                                isEditProductsRelease3Enabled: false)
-
-        // Assert
-        let expectedSettingsSectionActions: [ProductFormEditAction] = [.priceSettings]
-        XCTAssertEqual(factory.settingsSectionActions(), expectedSettingsSectionActions)
-
-        let expectedBottomSheetActions: [ProductFormBottomSheetAction] = [.editInventorySettings, .editBriefDescription]
-        XCTAssertEqual(factory.bottomSheetActions(), expectedBottomSheetActions)
-    }
-
-    // M3 feature flag on & M2 feature flag is on
-
     func testDataHasEditProductsRelease3ActionsForAPhysicalProductWhenBothFeatureFlagsAreOn() {
         // Arrange
         let product = Fixtures.physicalProduct
